added mouse interaction
Hovering over a tile with the mouse now highlights it via colour blending. Accuracy seems good but it doesn't feel great when the tiles are small, so if zooming becomes a thing I should probably disable it when zoomed out past a certain point. Also adjusted the widerMap function to better give the layout I was shooting for. Not sure if I acctually like it yet though so it might not stay.