Testing out noise with level gen
Might be that noise isnt the way to go with these small maps. Since Using a higher octave count means that its pretty likely the whole map is just solid or water, and a lower count looks really noisy and isnt a very good use of noise. A Solution to this could be some kind of edge detection where I algorithmically cycle through seeds until I find one has the ratio of water to ground that I specify. Also, for the jagged level function, co-ordinates go into the negative and the x and y in general is all weird. To use noise with that function, I would have to find the largest and smallest x and y co-ordinates, create a 2d noise array that is big enough to fit those tiles, then calculate some offset value when indexing the noise.