Slight map gen update, added label to menu button
Added little mountain sprites so the level isn't so flat. The code for drawing these takes up alot of space, since I dont really factor in a z level with the isometricToPixel function, so I should probably add that, as well as updating the map to support a z level and store the tile co-ordinates as a vector3 instead of vector2. Depends on how much verticality I plan to have, because if its just a base layer and top layer like it is now, updating to use Vec3 would be overkill, when I could instead just have 2 maps for base layer tiles and another data structure for the top layer that would store entities like mountains, units, etc.