Added camera zoom
Can now flick between 4 different zoom levels with I and O keys, which just changes the size the tiles are drawn at and updates the mouse co-ordinate calculation. Zooming out right now is a little disorienting since it sometimes completely changes the section of the map on screen. Not sure on the way to improve this, ideally, when zooming in and out it could account for the tile co-ordinate that is most centre to the screen and update the camera position such that this tile is still centred, but maybe theres a better way to do this