update map drawing to use layers
Gotta think about it some more, this solution would currently have problems with features like pathfinding, since the base layer doesnt store mountain info, I would need to constantly check if any 1 value tile also stored a value in the top layer, which could be needlessly slow. Easy fix there would be to store another map of 0's and 1's (traversable or not) which would be the collection used for pathfinding and such, but maybe theres an easier way to do all this. idk