Swapped out System.Random with custom prng
Replaced use of System.Random for generating random numbers with a custom class based on xoshiro256** (source code for that can be found here: https://prng.di.unimi.it/xoshiro256starstar.c). This way, it should give the same output regardless of machine and runtime and can be more easily serialized for when saving and loading is a thing.