World gen update
Added a some code after the water/mountain generation to use a floodfill algorithm to find sections of the map that were small enclosed amounts of water/land and remove them, this way there shouldnt be any small islands/bodies of water on the map, which could cause issue since the player spawns on a random tile on the map. That said there can still be continents and it's not guaranteed that every land tile is connected to every other land tile, so the player will still need a way to traverse water when I actually get around to some actual player movement code. Apart from this, I was testing out the pathfinding again and noticed that when the findPath function is ran from the player to an inaccessible tile (specifically a mountain or tree tile) that game noticeably lags. I probably need to add an explicit early return since I can easily figure out if a path is impossible before I start the calculations