Added better player movement and path find update
Player movement now works on worldTime, only moving every specified amount of in game minutes (could be changed to any amount though). Might add an animation to lerp between tiles, not sure how I feel about that yet though. Also made some updates to pathFinding. Before this, trying to pathfind from the player to an inaccessible tile (either because the tile was false in navmap or surrounded by falses) would cause lag, since it would end up checking every world tile before returning null. Now, I use the floodFill to swap out the navMap with a subset that only contains the accessible tiles. The solution is a bit hacky since I designed floodFill for the world Layer dictionaries, so I might go and change it or add method overloading later. Trees are still also counted as inaccessible on the navMap, mostly because im unsure of how I want to render the player tile while in one of these tiles, but that'll be changed eventually.