visible day night cycle added
worldTime class now stores a color Value  which is updated alongside time. Works by lerping 2 colors (based on if the time is between noon, dusk, dawn, or midnight) lerp accounts for time in hours and minutes. Seems to work great, colors not final though. Also added small part to worldLogic to add some structures to the map (not interactable yet). Eventually I want to refactor that though, solution right now is a bit hacky since the GetRandomPos function works on the navMap, since I designed it only for spawning units. This means I need to generate the world, use the world to generate the navMap, then pass the topLayer and navMap to the function to add structures.