gsm update + misc
I now have the gamestate manager hold the ContentManager as a member, and I can now create the gameScene using an event that can be triggered from MenuScene. This way, the game isn't created at runtime, but when the menu button is pressed. Eventually, I could probably use a similar system for creating all scenes, since when I'm in the game scene, all the loaded data in Menu and Settings will just be taking up space which isnt needed.