Lots of misc
Added an actual worldClass. Still wanna keep this as light as possible if I can, but made sense to be a class as it stores so much stuff (Camera and player could maybe go in here as well to be fair). I also added more events, so now when a scene change occurs, a new scene is instantiated and all others are nulled, since they don't wont be doing anything and just take up space. Not sure. Haven't actually tested the memory so theres a chance that the scenes actually don't get cleaned up because I've been stupid with the event handlers or content pipeline or some such, so might need a clean up later if thats the case.