rollback 3d stuff + misc
dropped the 3d stuff for now cause of changing ideas. Also removed GameScene being a manager for subscenes, but that might come back depending on scope and whatnot. In terms of additions, mostly just added a tiletype enum, so the layer collections store tile type as text instead of number values for better readability.