UI updates
Added 2 main things, the popup system and textBlock. TextBlock is just a ui component that stores a textArea, which stores a collection of lines of text. This way If I want to display a large amount of text, I can use this text area, which will parse and seperate the string into multiple strings which can be drawn line by line. This way I can display multiple lines of text, as opposed to a single line which can often go off screen. The main use of this class for now is the popup system. This is a ui component which stores a text block and a collection of interactive buttons. This can be used on event triggers (e.g. the player enters a structure tile) so that the buttons can be used to perform actions such as changing scene and so forth. Still a lot of todos for it though.