Many things
The game scene is now really a scene to store a new subscene manager. The original game scene code is now in MapScene and this is what the GameScenes currentScene value will originally be. The popup that triggers when the player enters a structure tile now switches the scene to a new dungeon class. Also added a pause menu so I quit to menu via an actual button now instead of just tabbing. Also some misc changes such as textbutton supporting large and small fonts based on a bool in the constructor and changing the events setup in the gameStateManager to no longer use anonymous lambdas for events, which I heard meant that nulling scenes wasn't giving me all the performance I expected. Worth noting that I also created a GameContext class. This is similar to how I might use a globals class for save state. It will store information that is needed across subscenes, examples being inventory contents, player position, active quests, the world itself and so on.