New scene added (dungeon) + misc
Got rid of tiletype, since I didnt really know how to handle it for different scenes (e.g. a dungeon scene will have completely different tile types than the map and ints easy enough to work with anyhow. In terms of the dungeon stuff, I got the scene set up and clicking enter on the popup that triggers when moving to a structure tile now switches scene to the new dungeon one. Nothing in there for now, just a basic map generated with a basic random walk algorithm.